+++ packages/backend/src/server/api/endpoints/reversi/invitations.ts
@@ -7,10 +7,13 @@ import { Inject, Injectable } from '@nestjs/common';
 import { Endpoint } from '@/server/api/endpoint-base.js';
 import { DI } from '@/di-symbols.js';
 import { UserEntityService } from '@/core/entities/UserEntityService.js';
+import { RoleService } from '@/core/RoleService.js';
 import { ReversiService } from '@/core/ReversiService.js';
+import { asyncFilter } from '@/misc/prelude/array.js';
 
 export const meta = {
 	requireCredential: true,
+	requireRolePolicy: 'canPlayGames',
 
 	kind: 'read:account',
 
@@ -28,10 +31,14 @@ export const paramDef = {
 export default class extends Endpoint<typeof meta, typeof paramDef> { // eslint-disable-line import/no-default-export
 	constructor(
 		private userEntityService: UserEntityService,
+		private roleService: RoleService,
 		private reversiService: ReversiService,
 	) {
 		super(meta, paramDef, async (ps, me) => {
-			const invitations = await this.reversiService.getInvitations(me);
+			const invitations = await asyncFilter(await this.reversiService.getInvitations(me), async (userId) => {
+				const policies = await this.roleService.getUserPolicies(userId);
+				return policies.canPlayGames;
+			});
 
 			return await this.userEntityService.packMany(invitations, me);
 		});